+++ src/cobib/ui/components/__init__.py
@@ -15,12 +15,10 @@ from .entry_view import EntryView as EntryView
 from .help_screen import HelpScreen as HelpScreen
 from .input_screen import InputScreen as InputScreen
 from .list_view import ListView as ListView
+from .log_screen import LogScreen as LogScreen
+from .logging_handler import LoggingHandler as LoggingHandler
 from .main_content import MainContent as MainContent
 from .motion_key import MotionKey as MotionKey
-from .popup import Popup as Popup
-from .popup_logging_handler import PopupLoggingHandler as PopupLoggingHandler
-from .popup_panel import PopupPanel as PopupPanel
 from .preset_filter_screen import PresetFilterScreen as PresetFilterScreen
-from .prompt import Prompt as Prompt
 from .search_view import SearchView as SearchView
 from .selection_filter import SelectionFilter as SelectionFilter